requirements.txt
networkx==2.4
paho-mqtt==1.5.0
rhasspy-hermes==0.2.0
rhasspy-nlu==0.1.8
snips-nlu==0.20.2
